refactor: Move AnnounceUri to its own folder
