+++ lib/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod parser;
+pub mod tracker;
