+++ lib/src/parser/file.rs
@@ -1,4 +1,5 @@
-use super::{AnnounceUri, MetaInfo};
+use crate::{parser::MetaInfo, tracker::AnnounceUri};
+
 use anyhow::Result;
 use serde::{
    Deserialize, Serialize, Serializer,
